修复 qwen3omni zero2 梯度通信问题
