+++ src/llamafactory/model/patcher.py
@@ -17,8 +17,6 @@ from typing import TYPE_CHECKING, Any
 
 import torch
 from peft import PeftModel
-from torch import nn
-from torch.nn import functional as F
 from transformers import GenerationMixin, PreTrainedModel, PreTrainedTokenizerBase
 from transformers.integrations import is_deepspeed_zero3_enabled
 from transformers.modeling_utils import is_fsdp_enabled
@@ -52,71 +50,10 @@ if is_transformers_version_greater_than("4.57.0"):
 logger = logging.get_logger(__name__)
 
 
-class Qwen3OmniMoeThinkerTextSparseMoeBlock(nn.Module):
-    def __init__(self, config):
-        super().__init__()
-        self.num_experts = config.num_experts
-        self.top_k = config.num_experts_per_tok
-        self.norm_topk_prob = config.norm_topk_prob
-
-        # gating
-        self.gate = nn.Linear(config.hidden_size, config.num_experts, bias=False)
-        self.experts = nn.ModuleList(
-            [
-                modeling_qwen3_omni_moe.Qwen3OmniMoeThinkerTextMLP(
-                    config, intermediate_size=config.moe_intermediate_size
-                )
-                for _ in range(self.num_experts)
-            ]
-        )
-
-    def forward(self, hidden_states: torch.Tensor) -> torch.Tensor:
-        batch_size, sequence_length, hidden_dim = hidden_states.shape
-        hidden_states = hidden_states.view(-1, hidden_dim)
-        # router_logits: (batch * sequence_length, n_experts)
-        router_logits = self.gate(hidden_states)
-
-        # Calculate the routing weights for all experts
-        routing_weights = F.softmax(router_logits, dim=1, dtype=torch.float)
-
-        # Retain the weight of the top_k and reset the rest of the expert rights to 0 (instead of retaining only top_k experts)
-        top_k_weights, top_k_indices = torch.topk(routing_weights, self.top_k, dim=-1)
-        # Initialize the all-zero weight matrix (same shape as all experts)
-        full_routing_weights = torch.zeros_like(routing_weights)
-        # Only the weight of top_k experts is retained, and the weight of the rest of the experts remains at 0
-        full_routing_weights.scatter_(1, top_k_indices, top_k_weights)
-
-        # Normalized top_k weights (keep the original logic consistent)
-        if self.norm_topk_prob:
-            # Calculate the sum of the weights top_k each row (for normalization)
-            top_k_sum = full_routing_weights.sum(dim=-1, keepdim=True)
-            # Avoid dividing by zero
-            top_k_sum = torch.clamp(top_k_sum, min=1e-9)
-            full_routing_weights /= top_k_sum
-
-        # Convert back to the input data type
-        full_routing_weights = full_routing_weights.to(hidden_states.dtype)
-
-        final_hidden_states = torch.zeros(
-            (batch_size * sequence_length, hidden_dim), dtype=hidden_states.dtype, device=hidden_states.device
-        )
-
-        # Go through all the experts (not just the selected ones)
-        for expert_idx in range(self.num_experts):
-            expert_layer = self.experts[expert_idx]
-            # Get the weight of the current expert (inactive expert has a weight of 0 here)
-            expert_weights = full_routing_weights[:, expert_idx, None]  # shape: (batch*seq, 1)
-            # All samples participate in the calculations of the current expert, the weight may be equal to 0
-            current_hidden_states = expert_layer(hidden_states) * expert_weights
-            # Add-up to all expert outputs (experts with a weight of 0 do not affect the result)
-            final_hidden_states += current_hidden_states
-
-        final_hidden_states = final_hidden_states.reshape(batch_size, sequence_length, hidden_dim)
-        return final_hidden_states, router_logits
-
-
 def patch_qwen3_omni_moe_thinker_text_sparse_moe_block():
     if is_transformers_version_greater_than("4.57.0"):
+        from .model_utils.moe import Qwen3OmniMoeThinkerTextSparseMoeBlock
+
         modeling_qwen3_omni_moe.Qwen3OmniMoeThinkerTextSparseMoeBlock = Qwen3OmniMoeThinkerTextSparseMoeBlock
